sorted out some cyclic dependencies; further testing and debug is needed
