+++ cakechat/utils/s3/__init__.py
@@ -1,3 +1,2 @@
 from cakechat.utils.s3.bucket import S3Bucket
-from cakechat.utils.s3.utils import get_s3_resource, get_s3_model_resolver
-from cakechat.utils.s3.resolver import S3FileResolver
+from cakechat.utils.s3.resolver import S3FileResolver, get_s3_resource, get_s3_model_resolver
