+++ cakechat/utils/w2v/__init__.py
@@ -1,2 +1,1 @@
-from cakechat.utils.w2v.model import get_w2v_model
 from cakechat.utils.w2v.utils import get_w2v_model_path, get_w2v_params_str, get_w2v_model_name
